insert, delete, select using spring jdbc

+++ restful-web-services/src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 
 create table course (
-    id bigint not null,
+    id int not null,
     name varchar(255) not null,
     author varchar(255) not null,
     primary key (id)
